+++ pigit/config.py
@@ -4,9 +4,10 @@ from typing import Dict, List
 import os, re, textwrap, logging
 from distutils.util import strtobool
 
+from plenty.style import Color
+
 from .common.utils import confirm, traceback_info
 from .common.singleton import Singleton
-from .render.style import Color
 
 Logger = logging.getLogger(__name__)